pyproject.toml
[project]
name = "analysis-pension-ceilings"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Athroniaeth", email = "pierre.chaumont@hotmail.fr" }
]
dependencies = [
    "typer>=0.12.5",
    "python-dotenv>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/analysis_pension_ceilings"]
